Refactor ID types from size_t to long in various modules
- Updated function signatures in armes.h, bibelots.h, objets.h to use long instead of size_t for ID parameters.
- Modified implementations in armes.c, bibelots.c, objets.c to accommodate the new ID type.
- Adjusted print statements in display.c to reflect the change in ID type.
- Refactored bestiary initialization in creatures.c to handle groups and creatures with long IDs.
- Updated configuration loading functions to handle long IDs for creatures and groups.
- Ensured proper memory management for new structures in bestiary.
- Adjusted save/load functionality in sauvegarde.c to handle long IDs for equipped weapons.

+++ code/include/armes.h
@@ -8,8 +8,8 @@
 
 Arsenal *chargerArmesDepuisFichier(char  *filename);
 void afficherArmes(Arsenal *arsenal);
-int ajouterArme(Arsenal *modal, Arsenal *arsenal, size_t id_arme);
-int equiperArme(Plongeur *joueur, size_t id_arme);
+int ajouterArme(Arsenal *modal, Arsenal *arsenal, long id_arme);
+int equiperArme(Plongeur *joueur, long id_arme);
 int appliquerActionsArme(Plongeur *joueur, void *cible, EntiteType cible_type);
 
 void freeArme(Arme *arme);
